requirements.txt
fastapi
uvicorn
kafka-python
confluent-kafka
numpy
opencv-python
torch
torchvision
torchaudio
tensorflow
prometheus_client
redis
pymongo
psycopg2-binary
sqlalchemy
requests
pydantic
pydantic-settings
python-dotenv
pytest
motor
python-jose[cryptography]
passlib[bcrypt]
httpx
websockets
bcrypt
python-multipart
